+++ docker/entrypoint.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-DIR=/docker-entrypoint.d
-if [[ -d "$DIR" ]]
-then
-  /bin/run-parts  "$DIR"
-fi
-
-exec "$@"